+++ pyproject.toml
@@ -8,6 +8,8 @@ authors = ["matthew <matthew@cupist.com>"]
 python = "^3.7"
 flask = "^1.1.2"
 pytest = "^5.4.1"
+pytest-testmon = "^1.0.2"
+pytest-watch = "^4.2.0"
 
 [tool.poetry.dev-dependencies]
