+++ src/domain/__init__.py
@@ -1,1 +1,2 @@
 from .main import api_main
+from .user import api_user
