+++ src/domain/user/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint, jsonify
+
+api_user = Blueprint("for_user", __name__, url_prefix="/api/v1/user")
+
+from .user import *
