+++ src/domain/user/controller.py
@@ -1,0 +1,7 @@
+from .user import api_user, jsonify
+
+
+@api_user.route("/", methods=["GET", "PSOT"])
+def register():
+    return jsonify("success")
+
