+++ src/tests/basic/test_fixture.py
@@ -30,3 +30,16 @@ def test_sample(sample_db):
     print(sample_db)
     assert sample_db is "pg"
 
+
+@pytest.mark.parametrize(
+    "a, b",
+    [
+        (1, {"Two Scoops of Django": "1.8"}),
+        (True, "Into the Brambles"),
+        ("Jason likes cookies", [1, 2, 3]),
+        ("PYTEST_PLUGIN", "plugin_template"),
+    ],
+    ids=["int and dict", "bool and str", "str and list", "CookiecutterTemplate and str",],
+)
+def test_foobar(a, b):
+    assert True
