+++ src/tests/domain/test_user.py
@@ -1,0 +1,3 @@
+def test_post_user_should_success(flask_client):
+    res = flask_client.post("/api/v1/user")
+    assert res.status_code == 300
