+++ src/tests/test_main.py
@@ -5,5 +5,4 @@ import pytest
 @pytest.mark.parametrize("env", [("dev"), ("prod"), ("test")])
 def test_create_server_on_env(env):
     app = create_app(env)
-
     assert app.config["NAME"] == env.upper()
